LayoutTests/http/tests/security/local-video-src-from-remote-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: scaled-matrix.mov
Test that a remote video element can not use a local 'src'

This test only works in DRT

EVENT(error)

SUCCESS: local movie failed to load

END OF TEST
